# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Laravel Generator APIs",
        "version": "1.0.0"
    },
    "basePath": "/teachme-api/public/api",
    "paths": {
...
        "/disciplinaLecionas": {
            "get": {
                "tags": [
                    "Disciplina_Leciona"
                ],
                "summary": "Get a listing of the Disciplina_Lecionas.",
                "description": "Get all Disciplina_Lecionas",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Disciplina_Leciona"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Disciplina_Leciona"
                ],
                "summary": "Store a newly created Disciplina_Leciona in storage",
                "description": "Store Disciplina_Leciona",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Disciplina_Leciona that should be stored",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Disciplina_Leciona"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Disciplina_Leciona"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/disciplinaLecionas/{id}": {
            "get": {
                "tags": [
                    "Disciplina_Leciona"
                ],
                "summary": "Display the specified Disciplina_Leciona",
                "description": "Get Disciplina_Leciona",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Disciplina_Leciona",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Disciplina_Leciona"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Disciplina_Leciona"
                ],
                "summary": "Update the specified Disciplina_Leciona in storage",
                "description": "Update Disciplina_Leciona",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Disciplina_Leciona",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Disciplina_Leciona that should be updated",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Disciplina_Leciona"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Disciplina_Leciona"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Disciplina_Leciona"
                ],
                "summary": "Remove the specified Disciplina_Leciona from storage",
                "description": "Delete Disciplina_Leciona",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Disciplina_Leciona",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Disciplina_Leciona": {
            "required": [
                "CdDisciplina",
                "CdUsuario_Professor"
            ],
            "properties": {
                "CdDisciplina": {
                    "description": "CdDisciplina",
                    "type": "integer",
                    "format": "int32"
                },
                "CdUsuario_Professor": {
                    "description": "CdUsuario_Professor",
                    "type": "integer",
                    "format": "int32"
                }
            }
        },
...
    }
}